fixed bug with editing words
