+++ src/components/words/WordsPage.jsx
@@ -4,6 +4,12 @@ import { Adding } from "../form/Adding";
 import "bootstrap/dist/css/bootstrap.min.css";
 import apiQueries from "../../api/apiQueries";
 
+const addNewWord = (word) => {
+  if (word.eng && word.rus) {
+    apiQueries.addItem(word.eng.trim(), word.rus.trim());
+  }
+};
+
 export const WordsPage = () => {
   const [wordsList, setWordsList] = useState([]);
   const [userWord, setUserWord] = useState({ eng: "", rus: "" });
@@ -12,16 +18,9 @@ export const WordsPage = () => {
     apiQueries.getData(setWordsList);
   };
 
-  // TODO: check this function
   useEffect(() => {
     showList();
   }, []);
-
-  const addNewWord = (word) => {
-    if (word.eng && word.rus) {
-      apiQueries.addItem(word.eng.trim(), word.rus.trim(), wordsList); // TODO: delete wordsList and move beyond component
-    }
-  };
 
   const removeWord = (id) => {
     apiQueries.deleteItem(id);
